pyproject.toml
[project]
name = "ai-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq<1",
    "langchain>=1.2.18",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
]
